Replaced CoinDetailsSubtitle.kt and CoinDetailsText.kt by CoinDetailsTextBlock.kt
