--- features/src/main/kotlin/ru/mobileup/template/features/cryptocurrency/presentation/details/ui/CoinDetailsSubtitle.kt
@@ -1,18 +1,0 @@
-package ru.mobileup.template.features.cryptocurrency.presentation.details.ui
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import ru.mobileup.template.core.theme.custom.CustomTheme
-
-@Composable
-fun CoinDetailsSubtitle(
-    title: String,
-    modifier: Modifier = Modifier
-) {
-    Text(
-        text = title,
-        style = CustomTheme.typography.body.regular,
-        modifier = modifier
-    )
-}
